Check if file exists before attempting to overwrite it
Fixes: #3267

+++ rewrite-java/src/main/java/org/openrewrite/java/JavaParser.java
@@ -37,10 +37,7 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.net.URI;
 import java.nio.charset.Charset;
-import java.nio.file.FileAlreadyExistsException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.nio.file.*;
 import java.util.*;
 import java.util.function.Function;
 import java.util.regex.Matcher;
@@ -172,14 +169,14 @@ public interface JavaParser extends Parser<J.CompilationUnit> {
                     if (jarPattern.matcher(resource.getPath()).find()) {
                         try {
                             Path artifact = resourceTarget.toPath().resolve(Paths.get(resource.getPath()).getFileName());
-
-                            Files.copy(
-                                    requireNonNull(caller.getResourceAsStream("/" + resource.getPath())),
-                                    artifact
-                            );
+                            if (!Files.exists(artifact)) {
+                                Files.copy(
+                                        requireNonNull(caller.getResourceAsStream("/" + resource.getPath())),
+                                        artifact
+                                );
+                            }
                             missingArtifactNames.remove(artifactName);
                             artifacts.add(artifact);
-                        } catch (FileAlreadyExistsException ignore) {
                         } catch (IOException e) {
                             throw new UncheckedIOException(e);
                         }
